+++ tests/conftest.py
@@ -4,4 +4,4 @@ import sys
 import pkg_resources
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
-pkg_resources.require('wsgidav')
+pkg_resources.require("wsgidav")