+++ wsgidav/server/ext_wsgiutils_server.py
@@ -282,16 +282,16 @@ class ExtServer (socketserver.ThreadingMixIn, BaseHTTPServer.HTTPServer):
         # 10053: Software caused connection abort
         # 10054: Connection reset by peer
         if e.args[0] in (10053, 10054):
-            util.warn("*** Caught socket.error: {}".format(e))
+            _logger.error("*** Caught socket.error: {}".format(e))
             return
         # This is what BaseHTTPServer.HTTPServer.handle_error does, but with
         # added thread ID and using stderr
-        _logger.error('-' * 40, file=sys.stderr)
-        _logger.error('<{}> Exception happened during processing of request from {}'
+        _logger.error("-" * 40, file=sys.stderr)
+        _logger.error("<{}> Exception happened during processing of request from {}"
                       .format(threading.currentThread().ident, client_address))
         _logger.error(client_address, file=sys.stderr)
         traceback.print_exc()
-        _logger.error('-' * 40, file=sys.stderr)
+        _logger.error("-" * 40, file=sys.stderr)
         _logger.error(request, file=sys.stderr)
 #        BaseHTTPServer.HTTPServer.handle_error(self, request, client_address)