+++ wsgidav/server/run_server.py
@@ -70,6 +70,22 @@ def _get_checked_path(path, mustExist=True, allowNone=True):
     return path
 
 
+class FullExpandedPath(argparse.Action):
+    """Expand user- and relative-paths"""
+    def __call__(self, parser, namespace, values, option_string=None):
+        new_val = os.path.abspath(os.path.expanduser(values))
+        setattr(namespace, self.dest, new_val)
+
+
+# def arg_is_dir(dirname):
+#     """Checks if a path is an actual directory"""
+#     print("is_dir", dirname)
+#     if not os.path.isdir(dirname):
+#         msg = "{} is not a directory".format(dirname)
+#         raise argparse.ArgumentTypeError(msg)
+#     return dirname
+
+
 def _initCommandLineOptions():
     """Parse command line options into a dictionary."""
 
@@ -117,6 +133,8 @@ See https://github.com/mar10/wsgidav for additional information.
                               "application public")),
     parser.add_argument("-r", "--root",
                         dest="root_path",
+                        action=FullExpandedPath,
+                        # type=arg_is_dir,
                         help="path to a file system folder to publish as share '/'.")
     parser.add_argument("--server",
                         choices=("cheroot", "cherrypy-wsgiserver", "ext-wsgiutils", "flup-fcgi",
@@ -137,13 +155,14 @@ See https://github.com/mar10/wsgidav for additional information.
 
     parser.add_argument("-c", "--config",
                         dest="config_file",
+                        action=FullExpandedPath,
                         help=("configuration file (default: {} in current directory)"
                               .format(DEFAULT_CONFIG_FILES)))
     parser.add_argument("--no-config",
                         action="store_true", dest="no_config",
                         help="do not try to load default {}".format(DEFAULT_CONFIG_FILES))
 
-    parser.add_argument('-V', '--version', action='version', version=__version__)
+    parser.add_argument("-V", "--version", action="version", version=__version__)
 
     args = parser.parse_args()
 
@@ -154,8 +173,9 @@ See https://github.com/mar10/wsgidav for additional information.
         del args.quiet
     # print("Verbosity: {}".format(args.verbose))
 
-    if args.config_file and args.no_config:
-        parser.error("--config and --no-config are mutually exclusive")
+    if not os.path.isdir(args.root_path):
+        msg = "{} is not a directory".format(args.root_path)
+        raise parser.error(msg)
 
     if args.no_config:
         if args.config_file:
@@ -577,7 +597,10 @@ def run():
                          }
     config = _initConfig()
 
+    util.initLogging(config["verbose"], config.get("enable_loggers", []))
+
     app = WsgiDAVApp(config)
+
     server = config["server"]
     handler = SUPPORTED_SERVERS.get(server)
     if not handler: